+++ Apizr/Apizr/Configuring/Proper/ApizrProperOptionsBuilder.cs
@@ -75,6 +75,14 @@ namespace Apizr.Configuring.Proper
         }
 
         /// <inheritdoc />
+        public IApizrProperOptionsBuilder WithHttpClient(Func<HttpMessageHandler, Uri, HttpClient> httpClientFactory)
+        {
+            Options.HttpClientFactory = httpClientFactory;
+
+            return this;
+        }
+
+        /// <inheritdoc />
         public IApizrProperOptionsBuilder WithAuthenticationHandler(Func<HttpRequestMessage, Task<string>> refreshTokenFactory)
         {
             var authenticationHandler = new Func<ILogger, IApizrOptionsBase, DelegatingHandler>((logger, options) =>