+++ Apizr/Apizr/Configuring/Shared/IApizrSharedOptionsBuilder.cs
@@ -45,6 +45,13 @@ namespace Apizr.Configuring.Shared
         TApizrSharedOptionsBuilder WithHttpClientHandler(Func<HttpClientHandler> httpClientHandlerFactory);
 
         /// <summary>
+        /// Provide a custom HttpClient
+        /// </summary>
+        /// <param name="httpClientFactory">An <see cref="HttpClient"/> instance factory</param>
+        /// <returns></returns>
+        TApizrSharedOptionsBuilder WithHttpClient(Func<HttpMessageHandler, Uri, HttpClient> httpClientFactory);
+
+        /// <summary>
         /// Provide your own <see cref="AuthenticationHandlerBase"/> implementation factory
         /// </summary>
         /// <typeparam name="TAuthenticationHandler">Your <see cref="AuthenticationHandlerBase"/> implementation</typeparam>